debian/ubuntu-snippet with sudo
I generally like when snippets are provided in a way which could be used without running as root, and uses sudo when applicable. This change allows for this. 

It will, however print out one extra line, which is possible to remove by adding '>/dev/null' after '/etc/apt/sources.list.d/zrepl.list'.