By configuring your push or pull job with:
```yaml
conflict_resolution:
  initial_replication:
    replicate_all_snapshots: true
```

when a dataset is transfered for the first time, all snapshots on
that dataset will be transfered. This is disabled by default, and
only the most recent snapshot will be transfered.

Based on the names in https://github.com/zrepl/zrepl/issues/408#issuecomment-795329846
though the more complicated options suggested in that issue are
more Go than I know.