Fixes #410 - Uses zstream token.
Removed resumeTokenContentsRE and accompanying code (As zstream token doesn't have a starting string to regex against).

Updated resumeTokenNVListRE to match the new format for zstream token.

Updated some comments.

Actually use zstream token.

Commented out some of the other optional fields this returns as they are never in the actual token: This may be added back if later on we use zfs send -nvt on the sender side.

This is a quick and dirty fix, but it does work on my system to resume stuck transfers (And thus also fixes #510 albeit not using the best solution)