cover the condition in tests
The new test `ReplicationIncrementalHandlesFromVersionEqTentativeCursorCorrectly` fails with

	panic: impl error: "#zrepl_CURSORTENTATIVE__G_15cbdccbc1f99b2f_J_sender-job" would be destroyed because it is considered stale but it is part of of sendArgs=zfs.ZFSSendArgsValidated

The updated existing test `ReplicationIncrementalCleansUpStaleAbstractionsWithCacheOnSecondReplication`, which, despite what its name suggests, invalidates the cache, fails with

	panic: impl error: "@3" would be destroyed because it is considered stale but it is part of of sendArgs=zfs.ZFSSendArgsValidated

The test that doesn't invalidate the cache doesn't fail.
That makes sense because without the cursor in the cache, we can't make the overly cautious decision that causes the panic.

 with '#' will be ignored, and an empty message aborts the commit.