docs: switch to sphinx-multiversion for multi-versioned docs
The sphinxcontrib-versioning seems unmaintainted and I can't
get the fork that we used before this PR working on Python 3.10.

The situation wrt maintenance doesn't seem much better for
sphinx-multiversion, but, at least I could get it to work
with current sphinx versions.

The main drawback of this PR is that we're breaking the default URLs.
I.e., `https://zrepl.github.io/configuration.html` will be 404.
And since we're on GitHub pages and don't control the webserver,
we can't install a redirect easily.

The above is not necessarily a show-stopper, but not great either.
So, not merging this PR right away but looking into `sphinx-polyversion`
as recommended in https://github.com/Holzhaus/sphinx-multiversion/issues/88#issuecomment-1606221194

Notes
=====

The `requirements.txt` in this PR is from a clean Python 3.10 venv on
Ubuntu 22.10 after

```
pip install sphinx sphinx-rtd-theme sphinx-multiversion
```