Allow same root_fs for different jobs: sinks and so on
The idea of this check is
```
// idea:
//   no path in rfss must be prefix of another
```

Technically, with two equal root_fs, one of them is really prefix of other. But
logically, and I think we need exactly this, a prefix is a part of some string.
So two equal root_fs are not a prefix of each other.

And I don't see any reason why we can't have jobs with same root_fs. Anyway we
are root_fs + client's name, so names of datasets never are never overlapping.