Serve Prometheus metrics over (m)TLS
Add a TLSConfig struct/library to configure a simple TLS/HTTPS server
with most basic toggles that most users would want. This is based on the
structure that the prometheus/exporter-toolkit web-config uses.

Notably this does not make use of either of the two existing TLS
configuration structures that exist as both are specific in their usage
and neither are suitable for generally configuring a TLS server.

Signed-off-by: Joe Groocock <me@frebib.net>